[add]ユーザ情報の管理
